shifting is removed. Isothermal bc calculations are added.

+++ problems/immersed_boundaries/Cylinder2D.cpp
@@ -66,9 +66,10 @@ ImmersedBoundaries::setImmersedBoundaryVariablesOnPatch(
     /*
      * These will be read from the input file.
      */
+    
     const double radius_c = half; //double(20); AFK 
     const double x_c = double(5); //half;  AFK
-    const double y_c = double(4); //half;  AFK
+    const double y_c = double(5); //half;  AFK
     
     for (int j = domain_lo_1; j < domain_lo_1 + domain_dim_1; j++)
     {
